VDIVolumeManager, VolumeFromFile, AdaptiveVDIGenerator: refactors to modularize volume manager generation
